requirements.txt
# ── Core LLM / RAG framework ──────────────────────────────────────────────────
langchain==1.1.3
langchain-core==1.1.2
langchain-community==0.4.1
langchain-classic==1.0.0
langchain-groq==1.0.0
langchain-huggingface==1.0.1
langchain-text-splitters==1.0.0

# ── Embeddings & Reranking ────────────────────────────────────────────────────
sentence-transformers==5.0.0

# ── Vector store (Qdrant — open source, Apache 2.0) ──────────────────────────
qdrant-client==1.17.1
langchain-qdrant==1.1.0

# ── Chat memory (Redis — open source, BSD license; optional) ─────────────────
redis==7.1.0

# ── PDF parsing (needed only for ingestion) ───────────────────────────────────
llama-parse
llama-index-core
llama-index-readers-file

# ── LLM provider ─────────────────────────────────────────────────────────────
groq==0.30.0

# ── Web framework ────────────────────────────────────────────────────────────
fastapi==0.121.2
uvicorn[standard]==0.35.0
pydantic==2.11.7

# ── Security & middleware ─────────────────────────────────────────────────────
slowapi>=0.1.9
python-dotenv==1.0.1
